feat: removed all errors that were inConsole
