--- src/components/styles/EmployersTable.scss
@@ -1,49 +1,0 @@
-@import url("https://fonts.googleapis.com/css2?family=Fredoka+One&display=swap");
-@import url("https://fonts.googleapis.com/css2?family=Poppins:ital,wght@1,600&display=swap");
-
-* {
-  padding: 0;
-  margin: 0;
-  box-sizing: border-box;
-}
-h1 {
-  padding: 0.1em;
-  margin-top: -7em;
-  margin-bottom: 0.8em;
-  color:brown;
-}
-table {
- 
-  color: rgb(43, 12, 33);
-  width: 100%;
-
-
-  tbody {
-    font-weight: lighter;
-    font-family: "Poppins", sans-serif;
-    text-align: center;
-  }
-
-  tr{
-     display: flex;
-  justify-content: center;
-  align-items: center;
-  text-align: center;
-  }
-  
-  td {
-  border: none;
-  color: brown;
-  padding: 5px;
-  text-align: center;
-  
-  
-}
-
-
-}
-
-.image1{
-  width: 40px;
-  height: 40px;
-}
