+++ src/styles/App.scss
@@ -7,7 +7,7 @@
 }
 
 body {
-  background-image: url(./components/assets/Images/wallpaper.jpg);
+  background-image: url(../components/assets/Images/wallpaper.jpg);
   background-size: cover;
   background-repeat: no-repeat;
   width: 100%;
