+++ src/styles/EmployersTable.scss
@@ -5,50 +5,42 @@
   padding: 0;
   margin: 0;
   box-sizing: border-box;
-  
 }
-
-body{
+body {
   width: 100%;
   height: 100vh;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  
-  
-  h1 {
+}
+
+h1 {
   padding: 0.1em;
-  margin-top: -8em;
   margin-bottom: 0.8em;
   color: brown;
 }
 table {
   color: rgb(43, 12, 33);
-  
-
-
-  thead {
-    background-color: firebrick;
-    td {
-      color: white;
-
-    }
-  }
 
   tbody {
     font-weight: lighter;
     font-family: "Poppins", sans-serif;
+    text-align: center;
+  }
 
-    tr {
-      border: none;
-      border-top: 1px solid gray;
+  tr {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    text-align: center;
+  }
 
-    }
-    
+  td {
+    border: none;
+    color: brown;
+    padding: 5px;
+    text-align: center;
   }
 }
 
-
-
+.image1 {
+  width: 40px;
+  height: 40px;
 }
-
