+++ src/styles/Home.scss
@@ -8,7 +8,7 @@
 
 body {
   font-family: "Poppins", sans-serif;
-  background-image: url(../assets/Images/wallpaper.jpg);
+  background-image: url(../components/assets/Images/wallpaper.jpg);
   display: flex;
   justify-content: center;
   align-items: center;
