+++ src/styles/Navbar1.scss
@@ -8,9 +8,10 @@ body{
   display: flex;
   justify-content: space-between;
   align-items: center;
+  width: 150%;
 
   .logo {
-    margin-left: 0;
+    margin-left: -250px;
     color: white;
     font-size: 1em;
   }
@@ -28,6 +29,7 @@ body{
 
         a {
             text-decoration: none;
+            font-size: 1.5em;
         }
     }
   }
