+++ src/styles/ShoesTable.scss
@@ -1,0 +1,51 @@
+@import url("https://fonts.googleapis.com/css2?family=Fredoka+One&display=swap");
+@import url("https://fonts.googleapis.com/css2?family=Poppins:ital,wght@1,600&display=swap");
+
+* {
+  padding: 0;
+  margin: 0;
+  box-sizing: border-box;
+}
+
+body {
+  width: 100%;
+  height: 100vh;
+  
+ 
+  h1 {
+    padding: 0.1em;
+    margin-top: 4em;
+    margin-bottom: 0.8em;
+    color: brown;
+  }
+  table {
+    color: rgb(43, 12, 33);
+    width: 100%;
+    margin-bottom: 5em;
+    
+
+    thead {
+      background-color: brown;
+      width: 100%;
+      td {
+        color: white;
+        width: 100%;
+        border: rgb(43, 3, 3) solid 0.5px;
+      }
+    }
+
+    tbody {
+      font-weight: lighter;
+      font-family: "Poppins", sans-serif;
+
+      tr {
+        border: none;
+        border-top: 1px solid gray;
+        td {
+          width: 100%;
+          border: rgb(43, 3, 3) solid 0.5px;
+        }
+      }
+    }
+  }
+}
